hacking around getting the config for aggregates in, this might change
